# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Book Vault API
  version: 1.0.1
  description: |
    RESTful API for Book Vault audiobook library.

    ## Authentication

    **IMPORTANT**: This API supports **dual authentication** for maximum flexibility.
    All endpoints accept BOTH session cookies (web) AND Bearer tokens (mobile).

    ### Web Authentication (Session-based)
    - Login via NextAuth.js at `/api/auth/signin`
    - Session maintained via HTTP-only cookies
    - All documented endpoints support session auth

    ### Mobile Authentication (Bearer tokens)
    - Mobile-specific endpoints at `/api/auth/mobile/*`
    - Use JWT Bearer tokens: `Authorization: Bearer {token}`
    - All endpoints support both auth methods
    - See "Mobile Authentication" tag for mobile-specific endpoints

    ## Base URL
    - Development: http://localhost:3000
    - Production: TBD (AWS deployment)
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.bookvault.com
    description: Production server (TBD)
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie (HTTP-only)
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Mobile JWT token (limited endpoint support)
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
          example: 'Book not found'
    Pagination:
      type: object
      required: [page, limit, total, pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          minimum: 0
          example: 150
        pages:
          type: integer
          minimum: 0
          example: 8
          description: Total number of pages

    # Core models (from lib/types.ts)
    Book:
      type: object
      required: [id, asin, title, runtimeMinutes, coverUrl, audioUrl, authors]
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        asin:
          type: string
          example: 'B002V5D1CG'
        title:
          type: string
          example: 'The Name of the Wind'
        description:
          type: string
          nullable: true
          example: "Told in Kvothe's own voice, this is the tale of the magically gifted young man..."
        runtimeMinutes:
          type: integer
          minimum: 0
          example: 1068
        releaseDate:
          type: string
          format: date
          nullable: true
          example: '2008-04-01'
        publisher:
          type: string
          nullable: true
          example: 'DAW'
        coverUrl:
          type: string
          format: uri
          example: 'https://book-vault-media.s3.amazonaws.com/covers/B002V5D1CG.jpg'
        audioUrl:
          type: string
          format: uri
          example: 'https://book-vault-media.s3.amazonaws.com/audio/B002V5D1CG.m4b'
        authors:
          type: array
          items:
            $ref: '#/components/schemas/Author'
        narrators:
          type: array
          items:
            $ref: '#/components/schemas/Narrator'
        series:
          type: array
          items:
            $ref: '#/components/schemas/SeriesInfo'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
    Author:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Patrick Rothfuss'
        asin:
          type: string
          nullable: true
          example: 'B002BMAYSQ'
    Narrator:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Nick Podehl'
        asin:
          type: string
          nullable: true
    SeriesInfo:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: 'The Kingkiller Chronicle'
        sequence:
          type: string
          nullable: true
          description: "Book number in series (can be '1', '1.5', 'Prequel', etc.)"
          example: '1'
        asin:
          type: string
          nullable: true
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'Fantasy'
...
security:
  - sessionAuth: []
paths:
...
  /api/progress:
    get:
      tags: [Progress]
      summary: Get user progress for a book
      parameters:
        - name: bookId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Book ID to get progress for
      responses:
        '200':
          description: User progress
          content:
            application/json:
              schema:
                type: object
                required: [positionSeconds, completed]
                properties:
                  positionSeconds:
                    type: number
                    format: double
                    minimum: 0
                    example: 3612.5
                  completed:
                    type: boolean
                    example: false
                  lastPlayed:
                    type: string
                    format: date-time
                    nullable: true
                    example: '2025-12-25T10:30:00Z'
        '400':
          description: Missing book ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Progress]
      summary: Update user progress (save position)
      description: |
        **Supports both authentication methods**:
        - Session cookies (web)
        - Bearer token (mobile)
      security:
        - sessionAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, positionSeconds]
              properties:
                bookId:
                  type: string
                  format: uuid
                positionSeconds:
                  type: number
                  format: double
                  minimum: 0
                timestamp:
                  type: string
                  format: date-time
                  description: Optional client timestamp for conflict resolution
      responses:
        '200':
          description: Progress updated
          content:
            application/json:
              schema:
                type: object
                required: [positionSeconds, completed, lastPlayed, updated]
                properties:
                  positionSeconds:
                    type: number
                    format: double
                    minimum: 0
                    example: 3612.5
                  completed:
                    type: boolean
                    example: false
                  lastPlayed:
                    type: string
                    format: date-time
                    example: '2025-12-25T10:30:00Z'
                  updated:
                    type: boolean
                    description: Whether update was applied (false if timestamp conflict)
                    example: true

    put:
      tags: [Progress]
      summary: Mark book as completed or reset to not started
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId, status]
              properties:
                bookId:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [completed, not-started]
                  description: Mark as completed or reset progress
      responses:
        '200':
          description: Progress status updated
          content:
            application/json:
              schema:
                type: object
                required: [positionSeconds, completed]
                properties:
                  positionSeconds:
                    type: number
                    format: double
                    minimum: 0
                    example: 0
                  completed:
                    type: boolean
                    example: true
                  lastPlayed:
                    type: string
                    format: date-time
                    nullable: true
                    example: '2025-12-25T10:30:00Z'

  # Search endpoints
...
  /api/narrators/{id}:
    get:
      tags: [Browse]
      summary: Get narrator details and their books
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Narrator ID
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Narrator with their books
          content:
            application/json:
              schema:
                type: object
                required: [id, name, books, pagination]
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                    example: 'Nick Podehl'
                  asin:
                    type: string
                    nullable: true
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '404':
          description: Narrator not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/library:
    get:
      tags: [Library]
      summary: Get user's library books
      responses:
        '200':
          description: User's library
          content:
            application/json:
              schema:
                type: object
                required: [books, total]
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  total:
                    type: integer
                    minimum: 0
                    example: 42
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Library]
      summary: Add book to user's library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Book added to library
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
                    example: 'Book added to library'
        '200':
          description: Book already in library
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    type: string
                    example: 'Book already in library'
        '400':
          description: Missing book ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...